+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes to Calva.
 
 ## [Unreleased]
+- [Add built-in REPL Connect Sequences for Plain CLJS Browser and Node REPLs](https://github.com/BetterThanTomorrow/calva/issues/1114)
 
 ## [2.0.186] - 2021-04-10
 - [Allow keybindings to target when the cursor is inside a comment or a string](https://github.com/BetterThanTomorrow/calva/issues/1023)